Refactor transform blocks to use BaseBlock and dedicated module
- Migrated FlattenColumnsBlock, SelectorBlock, and SetToMajorityValue from utilblocks.py to blocks/transform/
- Enhanced blocks with BaseBlock inheritance for improved validation and logging
- Updated BlockRegistry usage with category and description parameters
- Maintained backward compatibility through legacy constructor support
- Fixed test expectations for improved error handling (MissingColumnError vs KeyError)
- Updated import paths and module structure
